+++ src/generator/annotations.ts
@@ -2,6 +2,7 @@ export const DTO_IGNORE_MODEL = /@DtoIgnoreModel/;
 export const DTO_READ_ONLY = /@DtoReadOnly/;
 export const DTO_ENTITY_HIDDEN = /@DtoEntityHidden/;
 export const DTO_CREATE_OPTIONAL = /@DtoCreateOptional/;
+export const DTO_CREATE_HIDDEN = /@DtoCreateHidden/;
 export const DTO_UPDATE_HIDDEN = /@DtoUpdateHidden/;
 export const DTO_API_CREATE_HIDDEN = /@DtoApiCreateHidden/;
 export const DTO_API_UPDATE_HIDDEN = /@DtoApiUpdateHidden/;