+++ src/generator/compute-model-params/compute-create-dto-params.ts
@@ -2,6 +2,7 @@ import slash from 'slash';
 import path from 'node:path';
 import {
   DTO_CREATE_OPTIONAL,
+  DTO_CREATE_HIDDEN,
   DTO_API_CREATE_HIDDEN,
   DTO_RELATION_CAN_CONNECT_ON_CREATE,
   DTO_RELATION_CAN_CREATE_ON_CREATE,
@@ -78,6 +79,7 @@ export const computeCreateDtoParams = ({
     )
       field.isReadOnly = false;
 
+    if (isAnnotatedWith(field, DTO_CREATE_HIDDEN)) return result;
     if (isReadOnly(field)) return result;
     if (isRelation(field)) {
       if (!isAnnotatedWithOneOf(field, DTO_RELATION_MODIFIERS_ON_CREATE)) {