+++ src/generator/compute-model-params/compute-update-dto-params.ts
@@ -7,7 +7,7 @@ import {
   DTO_RELATION_MODIFIERS_ON_UPDATE,
   DTO_TYPE_FULL_UPDATE,
   DTO_UPDATE_HIDDEN,
-  DTO_API_UPDATE_HIDDEN,
+  DTO_UPDATE_API_RESP,
   DTO_UPDATE_OPTIONAL,
 } from '../annotations';
 import {
@@ -55,6 +55,7 @@ export const computeUpdateDtoParams = ({
   templateHelpers,
 }: ComputeUpdateDtoParamsParam): UpdateDtoParams => {
   let hasApiProperty = false;
+  let hasApiRespProperty = false;
   const imports: ImportStatementParams[] = [];
   const extraClasses: string[] = [];
   const apiExtraModels: string[] = [];
@@ -68,7 +69,7 @@ export const computeUpdateDtoParams = ({
     const overrides: Partial<DMMF.Field> = {
       isRequired: false,
       isNullable: !field.isRequired,
-      updateApiHide: false,
+      updateApiResp: false,
     };
     const decorators: {
       apiProperties?: IApiProperty[];
@@ -192,7 +193,8 @@ export const computeUpdateDtoParams = ({
     }
 
     if (!templateHelpers.config.noDependencies) {
-      overrides.updateApiHide = isAnnotatedWith(field, DTO_API_UPDATE_HIDDEN);
+      overrides.updateApiResp = isAnnotatedWith(field, DTO_UPDATE_API_RESP);
+      hasApiRespProperty = hasApiRespProperty || overrides.updateApiResp;
       decorators.apiProperties = parseApiProperty(
         {
           ...field,
@@ -220,10 +222,11 @@ export const computeUpdateDtoParams = ({
     return [...result, mapDMMFToParsedField(field, overrides, decorators)];
   }, [] as ParsedField[]);
 
-  if (apiExtraModels.length || hasApiProperty) {
+  if (apiExtraModels.length || hasApiProperty || hasApiRespProperty) {
     const destruct = [];
     if (apiExtraModels.length) destruct.push('ApiExtraModels');
     if (hasApiProperty) destruct.push('ApiProperty');
+    if (hasApiRespProperty) destruct.push('ApiResponseProperty');
     imports.unshift({ from: '@nestjs/swagger', destruct });
   }